+++ lihil/interface/__init__.py
@@ -4,7 +4,10 @@ from typing import Any, Callable, Literal
 from typing import Protocol as Protocol
 from typing import TypeGuard, Union, get_args
 
+from msgspec import UNSET, Meta
 from msgspec import Struct as Struct
+from msgspec import UnsetType
+from msgspec import field
 from msgspec import field as field
 
 from lihil.interface.asgi import HTTP_METHODS as HTTP_METHODS
@@ -66,6 +69,13 @@ class _Missed:
 
 
 MISSING = _Missed()
+
+
+type Unset[T] = UnsetType | T
+
+
+def is_set[T](val: Unset[T]) -> TypeGuard[T]:
+    return val is not UNSET
 
 
 class RequestParamBase[T](Base):